Rename some functions for clarity.
