+++ retrochatbot/framework/domain/usecases/load_bot.py
@@ -1,10 +1,12 @@
 import importlib
 
+from retrochatbot.botapi.bot_api import Bot
+
 
 def load_bot(
     full_bot_class_name: str,
     participant_name: str,
-):
+) -> Bot:
     module_name, class_name = full_bot_class_name.rsplit(".", 1)
 
     bot_module = importlib.import_module(module_name)
